Convert the StableVec class to use virtual addressing (#3608)
* Converted StableVec to use u64 virtual addresses rather than physical

* Removed an extraneous comment

* Cleaned up the use statement

* Update stable_vec.rs to fix a non-building test after field rename

+++ program-runtime/src/invoke_context.rs
@@ -333,7 +333,7 @@ impl<'a> InvokeContext<'a> {
         // but performed on a very small slice and requires no heap allocations.
         let instruction_context = self.transaction_context.get_current_instruction_context()?;
         let mut deduplicated_instruction_accounts: Vec<InstructionAccount> = Vec::new();
-        let mut duplicate_indicies = Vec::with_capacity(instruction.accounts.len());
+        let mut duplicate_indicies = Vec::with_capacity(instruction.accounts.len() as usize);
         for (instruction_account_index, account_meta) in instruction.accounts.iter().enumerate() {
             let index_in_transaction = self
                 .transaction_context
